[refactor] (planner) split statistic package to stats, collection and derivation
