+++ fe/fe-core/src/main/java/org/apache/doris/catalog/Catalog.java
@@ -228,10 +228,10 @@ import org.apache.doris.qe.JournalObservable;
 import org.apache.doris.qe.VariableMgr;
 import org.apache.doris.resource.Tag;
 import org.apache.doris.service.FrontendOptions;
-import org.apache.doris.statistics.StatisticsJobManager;
-import org.apache.doris.statistics.StatisticsJobScheduler;
 import org.apache.doris.statistics.StatisticsManager;
-import org.apache.doris.statistics.StatisticsTaskScheduler;
+import org.apache.doris.statistics.collection.StatisticsJobManager;
+import org.apache.doris.statistics.collection.StatisticsJobScheduler;
+import org.apache.doris.statistics.collection.StatisticsTaskScheduler;
 import org.apache.doris.system.Backend;
 import org.apache.doris.system.Backend.BackendState;
 import org.apache.doris.system.Frontend;
