+++ fe/fe-core/src/main/java/org/apache/doris/planner/OlapScanNode.java
@@ -53,7 +53,7 @@ import org.apache.doris.common.util.Util;
 import org.apache.doris.qe.ConnectContext;
 import org.apache.doris.qe.SessionVariable;
 import org.apache.doris.resource.Tag;
-import org.apache.doris.statistics.StatsRecursiveDerive;
+import org.apache.doris.statistics.derivation.StatsRecursiveDerive;
 import org.apache.doris.system.Backend;
 import org.apache.doris.thrift.TExplainLevel;
 import org.apache.doris.thrift.TNetworkAddress;
