+++ fe/fe-core/src/main/java/org/apache/doris/statistics/collection/MetaStatisticsTask.java
@@ -15,7 +15,11 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.statistics;
+package org.apache.doris.statistics.collection;
+
+import org.apache.doris.statistics.StatsCategoryDesc;
+import org.apache.doris.statistics.StatsGranularityDesc;
+import org.apache.doris.statistics.StatsType;
 
 import java.util.List;
