+++ fe/fe-core/src/main/java/org/apache/doris/statistics/collection/StatisticsTask.java
@@ -15,10 +15,13 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.statistics;
+package org.apache.doris.statistics.collection;
 
 import org.apache.doris.catalog.Catalog;
 import org.apache.doris.common.DdlException;
+import org.apache.doris.statistics.StatsCategoryDesc;
+import org.apache.doris.statistics.StatsGranularityDesc;
+import org.apache.doris.statistics.StatsType;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
