+++ fe/fe-core/src/main/java/org/apache/doris/statistics/derivation/StatsRecursiveDerive.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.statistics;
+package org.apache.doris.statistics.derivation;
 
 import org.apache.doris.common.UserException;
 import org.apache.doris.planner.PlanNode;
